requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Google Cloud
google-cloud-bigquery==3.14.0
google-cloud-core==2.4.1
google-auth==2.25.0
google-cloud-logging==3.9.0

# Future ML capabilities
# google-cloud-aiplatform==1.38.0  # Uncomment when adding Vertex AI

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
